Replace DownloadFormListTask to coroutines
